Rename SinglePageDonationForm back to DonationForm
- there is only 1 donation form for the donation page supported currently,
so it can keep the basic name

https://phabricator.wikimedia.org/T368525
